requirements.txt
boto3==1.34.7
botocore==1.34.7
flask==3.0.0
redis==5.0.1
python-dateutil>=2.8.2
werkzeug>=3.0.1
requests>=2.31.0
gunicorn>=21.2.0
